add custom pin clipper
